requirements.txt
blinker==1.6.3
click==8.1.7
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
zipp==3.17.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
alembic==1.12.0
Flask-Migrate==4.0.5
flask-smorest==0.42.1
marshmallow==3.20.1
passlib==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7